Removed materials and drive types
Materials as a ctaegory is gone, it has been absorbed under "finish"
since we can only see the external features of the fasteners. We have:
- Plastic
- Brass
- Stainless Steel (Grey)
- Black Oxide Stainless Steel

We expect most of our fasteners to go into stainless steel (Grey), which
encompasses galvanized, zinc-planted, chrome-plated, etc. fasteners.

We also removed JIS because the difference between JIS and the more
common Phillips is very difficult to visually see