Implement multithreaded gdrive upload
UploadWorker is in charge of sending sessions to the cloud. It uploads
all session folders in the images directory.

"rclone copy" will not upload files added to a folder after the update
has occurred, which is good for our case since we don't want to upload
fastener imaging runs that are incomplete. (If someone starts an imaging
run while uploading)

We can access from the Samsung USB drive now.

Write to report right before upload to gdrive now.