implement two-shot-screw imaging
This utilizes the angle detection algorithm to send an instruction to the BP to rotate the imaging plane to the desired angle.

The BP will recognize the new command "start-two-shot-screw" and enter a mode that will accept "finished {number}" as an argument after the first image is taken. After that image is taken, it proceeds as usual, proceeding with a "finished" command. The BP will terminate after 2 images are taken.